+++ src/main/java/nl/siwoc/mediainfo/qtff/MdiaBox.java
@@ -18,7 +18,7 @@ package nl.siwoc.mediainfo.qtff;
 
 public class MdiaBox extends Box {
 	
-    public MdiaBox(Box parent, int size, byte[] data) throws Exception {
+    public MdiaBox(Box parent, long size, byte[] data) throws Exception {
 		setType("mdia");
 		setSize(size);
 		setParent(parent);
